add: dispersal logic #146
